Fix file misplacement
