+++ src/segmentedfilesystem/Main.java
@@ -1,15 +1,17 @@
 package segmentedfilesystem;
 
+import java.io.IOException;
+
 public class Main {
-    
+
     // If there's one command line argument, it is assumed to
     // be the server. If there are two, the second is assumed
     // to be the port to use.
-    public static void main(String[] args) {
+    public static void main(String[] args) throws IOException {
         String server = "localhost";
         // CHANGE THIS DEFAULT PORT TO THE PORT NUMBER PROVIDED
         // BY THE INSTRUCTOR.
-        int port = 0;
+        int port = 6014;
         
         if (args.length >= 1) {
             server = args[0];
