+++ src/segmentedfilesystem/PacketManager.java
@@ -1,3 +1,5 @@
+package segmentedfilesystem;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
